Ensure that the generate command terminates all plugins
This commit moves the signal handler and the top-level error recovery
into a utility function so that the same logic can be used by both
the apply/delete and the generate commands. This also makes it possible
to use the same cleanup for future commands.

Closes #134
